requirements.txt
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==4.0
django-allauth==0.47.0
django-ckeditor==6.2.0
@@ -19,10 +22,16 @@ django-jalali==5.0.0
django-js-asset==1.2.2
django-quill-editor==0.1.33
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fandogh-cli==1.26.10
idna==3.3
itypes==1.2.0
jdatetime==3.8.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
material==0.1
oauthlib==3.1.1
packaging==21.3
@@ -50,6 +59,7 @@ sorl-thumbnail==12.7.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.3
wrapt==1.13.3